+++ release_notes.md
@@ -3,6 +3,7 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-- Update Python Worker Version to [4.40.1](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.40.1)
+- Update Python Worker Version to [4.40.2](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.40.2)
 - Add JitTrace Files for v4.1044
 - Implementing a resolver that resolves worker configurations from specified probing paths (#11258)
+- Avoid emitting empty tag values for health check metrics (#11393)